Fix 3D coords for skyfallers and drop pods, add notifications on impact

+++ Source/Shields/Handlers/ActiveDropPodHandler.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using FrontierDevelopments.General;
 using Harmony;
 using RimWorld;
+using RimWorld.Planet;
 using UnityEngine;
 using Verse;
 
@@ -24,17 +25,23 @@ namespace FrontierDevelopments.Shields.Handlers
                 {
                     var pod = __instance;
                     if (pod?.Map == null) return true;
-                    Mod.ShieldManager.ImpactShield(pod.Map, Common.ToVector2(pod.Position), (shield, point) =>
+                    Mod.ShieldManager.ImpactShield(pod.Map, Common.ToVector3WithY(pod.Position, 0), (shield, point) =>
                     {
-                        if (shield.Damage(Mod.Settings.DropPodDamage, point))
+                        if (shield.IsActive())
                         {
-                            foreach (var pawn in pod.Contents.innerContainer.Where(p => p is Pawn))
+                            if (shield.Damage(Mod.Settings.DropPodDamage, point))
                             {
-                                pawn.Kill(new DamageInfo(new DamageDef(), 100));
+                                foreach (var pawn in pod.Contents.innerContainer.Where(p => p is Pawn))
+                                {
+                                    pawn.Kill(new DamageInfo(new DamageDef(), 100));
+                                }
+                                pod.Destroy();
+                                Messages.Message("fd.shields.incident.droppod.blocked.body".Translate(), new GlobalTargetInfo(pod.Position, pod.Map), MessageTypeDefOf.NeutralEvent);
+                                return true;
                             }
-                            pod.Destroy();
-                            return true;
+                            Messages.Message("fd.shields.incident.droppod.not_blocked.body".Translate(), new GlobalTargetInfo(pod.Position, pod.Map), MessageTypeDefOf.NegativeEvent);
                         }
+                        
                         return false;
                     });
                 }
